runtimes/Linux_armv7l/runtime/requirements.txt
pywasm3==0.5.0
fastapi==0.70.1
uvicorn==0.16.0
python-multipart
gpiozero
pynacl==1.4.0
